contributed to backend structure,routes,controllers and overall code transformation
