+++ backend/.env
@@ -1,0 +1,3 @@
+PORT=3000
+MONGODB_URI=
+ML_API_URL=http://localhost:8000
